search-service/requirements.txt
google-cloud-speech
langchain
# langchain-nvidia-ai-endpoints
langchain-chroma
langchain-cohere
flask
chromadb
# google-generativeai
flask-cors
python-dotenv
pip
setuptools
wheel
# spacy
rank_bm25
ipykernel
# langchain-huggingface